fix VERGEN_CARGO_FEATURES
The value of CARGO_FEATURE_* environment variable in build scripts is
"1". Thus a value of VERGEN_CARGO_FEATURES becomes something like
"1,1,1".

This commit fixes the 'is_cargo_feature' function to return the feature
name, as it appears in the CARGO_FEATURE_ environment variable name,
instead of the environment variable value.

Normalizing the name to lowercase seems safe, as feature names and
optional dependencies are usually named with lowercase letters.