Replace rustc commit hash/date when not present (#79)
Occasionally `rustc -vV` (used by `rustc_version`) will return `unknown`
for the commit date and/or hash. In `rustc_version::VersionMeta`, this
is represented as a `None` value in the `commit_hash` and `commit_date`
fields. This is a problem if these variables are depended on by the
application, as vergen will not panic if they are not present, but will
instead silently omit them.

This PR fixes this issue by replacing a `None` value for a commit hash
or date with the string "unknown".

Fixes #78