gix: support including untracked files in dirty check
gix does not provide an easy way to count untracked files in the dirty
check, so we have to do it manually. To do this we use a directory walk
set to emit untracked files, setting the dirty flag to true if any are
found. As currently configured I think it should ONLY emit untracked
files, but I'm not sure enough to bet on it. This potentially could be
expensive so we only perform this walk if dirty files were not already
detected.